Fix string representations of Transform class
The underlying C-array of floats is a 4x4 matrix, of which only a 3x3
subcomponent is relevant in the context of 2D transformations. The access of
its indices have been fixed accordingly. There was also a mix-up of col-major
and row-major order.

This fix makes the string representations consistent with the signature of
the from_values constructor. Tests ensuring this behaviour have been added.