Fix documentation build failure with sphinx 1.4
Sphinx 1.4 added a new warning which is emitted if it fails to syntax highlight
a block of code. This warning causes the documentation to fail to build with
this error because the code cannot be parsed as python:

  Warning, treated as error:
  /<builddir>/doc/source/examples.rst:69:
  WARNING: Could not lex literal_block as "python". Highlighting skipped.

This commit sets the correct syntax highlighting language so the code can be
parsed and the warning isn't emitted.